Fixed: type complains on Mac

+++ plugins/colorpicker/ColorFormatDlg.cpp
@@ -15,7 +15,7 @@ ColorFormatDlg::ColorFormatDlg( JuffPlugin* plugin, const QColor& color, QWidget
 	
 	_ui.btnHtml->setText( color.name() );
 	_ui.btnHex->setText( color.name().replace("#", "0x") );
-	_ui.btnHexSplitted->setText( QString().sprintf("0x%02hX, 0x%02hX, 0x%02hX", r, g, b) );
+	_ui.btnHexSplitted->setText( QString().sprintf("0x%02hX, 0x%02hX, 0x%02hX", (unsigned short)r, (unsigned short)g, (unsigned short)b) );
 	_ui.btnSplitted->setText( QString().sprintf("%i, %i, %i", r, g, b) );
 	
 	_ui.buttonGroup->setId( _ui.btnHtml, 0 );
